Create 287. Find the Duplicate Number.cpp
https://leetcode.com/problems/find-the-duplicate-number/description/

https://leetcode.com/problems/find-the-duplicate-number/solutions/1892921/9-approaches-count-hash-in-place-marked-sort-binary-search-bit-mask-fast-slow-pointers/

Binary Search
Note that the key is to find an integer in the array [1,2,…,n][1, 2, \dots, n][1,2,…,n] instead of finding an integer in the input array.

We can use the binary search algorithm, each round we guess one number, then scan the input array, narrow the search range, and finally get the answer.

According to the Pigeonhole Principle, n+1n + 1n+1 integers, placed in an array of length nnn, at least 111 integer will be repeated.

So guess a number first(the number midmidmid in the valid range [left,right][left, right][left,right]), count the elements of the array which is less than or equal to midmidmid in the array.

If cntcntcnt is strictly greater than midmidmid. According to the Pigeonhole Principle, repeated elements are in the interval [left,mid][left, mid][left,mid];
Otherwise, the repeated element is in the interval [mid+1,right][mid + 1, right][mid+1,right].
With extra O(1)O(1)O(1) space, without modifying the input.

Fast Slow Pointers
This problem is as same as 142. Linked List Cycle II, you can refer to this solution for the explanation of the slow fast pointer approach to solve this problem.

The key is to understand how to treat the input array as a linked list.

Take the array [1,3,4,2][1,3,4,2][1,3,4,2] as an example, the index of this array is [0,1,2,3][0, 1, 2, 3][0,1,2,3], we can map the index to the nums[n]nums[n]nums[n].

0→1→3→2→4→3→20 \to 1 \to 3 \to 2 \to 4 \to 3 \to 20→1→3→2→4→3→2

Start from nums[n]\textit{nums}[n]nums[n] as a new index, and so on, until the index exceeds the bounds. This produces a sequence similar to a linked list.

0→1→3→2→4→null0 \to 1 \to 3 \to 2 \to 4 \to null0→1→3→2→4→null

If there are a repeated numbers in the array, take the array [1,3,4,2,2][1,3,4,2,2][1,3,4,2,2] as an example,

0→1→3→2→4→3→2→4→20 \to 1 \to 3 \to 2 \to 4 \to 3 \to 2 \to 4 \to 20→1→3→2→4→3→2→4→2

Similarly, a linked list is like that:

0→1→3→2→4→2→4→2→…0 \to 1 \to 3 \to 2 \to 4 \to 2 \to 4 \to 2 \to \dots0→1→3→2→4→2→4→2→…

Here 2→42 \to 42→4 is a cycle, then this linked list can be abstracted as the following figure:

Link List

With extra O(n)O(n)O(n) space, without modifying the input.